requirements.txt
certifi==2023.7.22
charset-normalizer==3.0.1
click==8.1.3
defusedxml==0.7.1
h11==0.14.0
idna==3.7
jira==3.4.1
loguru==0.6.0
mypy==1.5.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pyaml==21.10.1
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
typing_extensions==4.6.0
urllib3==1.26.18
uvicorn==0.20.0
